Fix output newline issue
The last few numbers would appear on a new line.
This is because the calculated size values for the
output element's clone were being rounded,
resulting in an integer portion of a pixel
where the fraction portion is required to
properly display the output size.

I took out this rounding to correct the issue.